Android Application Source Files
The directory contains the source files of the android application through which the user's movement signals will be sent to the server and the user can upload the location history files to check his/her contacts with confirmed patients. Doctors can also upload the details of the past location of confirmed patients through this application.